requirements.txt
Django==1.5
PyJWT==0.2.1
South==1.0
django-crispy-forms==1.4.0
django-social-auth==0.8.1
oauthlib==0.6.3
psycopg2==2.5.4
python-social-auth==0.2.1
python3-openid==3.0.4
requests==2.4.1
requests-oauthlib==0.3.1
six==1.8.0
